refactor(Board.java): 클래스 분리
