+++ src/main/java/chess/domain/board/Board.java
@@ -6,7 +6,6 @@ import chess.domain.piece.PieceColor;
 import chess.domain.piece.PieceKind;
 import chess.domain.piece.strategy.MoveDirection;
 
-import java.util.EnumMap;
 import java.util.Map;
 
 public class Board {
@@ -23,8 +22,7 @@ public class Board {
     }
 
     public void move(Position source, Position target, PieceColor turnColor) {
-        checkMoving(source, target);
-        checkNoJumpPiece(source, target);
+        checkPath(source, target);
 
         Piece originalPiece = pieceAtPosition(source);
         originalPiece.movable(source, target);
@@ -33,11 +31,17 @@ public class Board {
         checkPawnCase(source, target, originalPiece);
         checkIsNotSameTeam(source, target);
 
-        judgeKingsState(target);
         movePiece(source, target, originalPiece);
     }
 
+    private void checkPath(Position source, Position target) {
+        checkMoving(source, target);
+        checkNoJumpPiece(source, target);
+    }
+
     private void movePiece(Position source, Position target, Piece originalPiece) {
+        judgeKingsState(target);
+
         putPieceAtPosition(target, originalPiece);
         putPieceAtPosition(source, VOID_PIECE);
     }
@@ -49,15 +53,15 @@ public class Board {
         }
     }
 
-    private void checkNoJumpPiece(Position source, Position target) {
-        if (source.isLineMove(target) || source.isDiagonalMove(target)) {
-            checkClearPath(source, target);
-        }
-    }
-
     private void checkMoving(Position source, Position target) {
         if (source.computeHorizontalDistance(target) == 0 && source.computeVerticalDistance(target) == 0) {
             throw new InvalidMoveException(STAY_ERROR_MESSAGE);
+        }
+    }
+
+    private void checkNoJumpPiece(Position source, Position target) {
+        if (source.isLineMove(target) || source.isDiagonalMove(target)) {
+            checkClearPath(source, target);
         }
     }
 
@@ -77,7 +81,6 @@ public class Board {
 
         Position pathPosition = new Position(source);
         pathPosition.moveUnit(xVector, yVector);
-
         while (!pathPosition.equals(target)) {
             checkIfClear(pathPosition);
             pathPosition.moveUnit(xVector, yVector);
@@ -126,49 +129,6 @@ public class Board {
         if (!piece.isSameColor(turnColor)) {
             throw new RuntimeException("해당 턴이 아닙니다.");
         }
-    }
-
-    public double computePoint(PieceColor pieceColor) {
-        double pointSum = 0.0;
-        for (XPosition xPosition : XPosition.values()) {
-            Map<PieceKind, Integer> existingPiece = countPieceKindAtColumn(pieceColor, xPosition);
-            pointSum += columnPoint(existingPiece);
-        }
-        return pointSum;
-    }
-
-    private Map<PieceKind, Integer> countPieceKindAtColumn(PieceColor pieceColor,
-                                             XPosition xPosition) {
-        Map<PieceKind, Integer> existingPiece = new EnumMap<>(PieceKind.class);
-        for (YPosition yPosition : YPosition.values()) {
-            Position currentPosition = Position.of(xPosition, yPosition);
-            Piece piece = pieceAtPosition(currentPosition);
-            checkSameColor(pieceColor, existingPiece, piece);
-        }
-
-        return existingPiece;
-    }
-
-    private void checkSameColor(PieceColor pieceColor, Map<PieceKind, Integer> existingPiece, Piece piece) {
-        if (piece.isSameColor(pieceColor)) {
-            existingPiece.put(piece.kind(), existingPiece.getOrDefault(piece.kind(), 0) + 1);
-        }
-    }
-
-    private double columnPoint(Map<PieceKind, Integer> column) {
-        double points = 0.0;
-        for (Map.Entry<PieceKind, Integer> entry : column.entrySet()) {
-            points += eachPiecePoint(entry);
-        }
-        return points;
-    }
-
-    private double eachPiecePoint(Map.Entry<PieceKind, Integer> entry) {
-        double point = entry.getValue() * entry.getKey().point();
-        if (entry.getKey() == PieceKind.PAWN && entry.getValue() > 1) {
-            point /= 2;
-        }
-        return point;
     }
 
     public Piece pieceAtPosition(Position position) {
